hello

+++ Test/Address.cs
@@ -8,7 +8,8 @@ namespace Test
         public string StreetAddress { get; set; }
         public string Suburb { get; set; }
         public string State { get; set; }
-        public decimal Latitude { get; set; }
-        public decimal Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
+     
     }
 }
